% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_caf_data.R
\name{caf2021_g2_groups}
\alias{caf2021_g2_groups}
\title{Default mapping of G2 codes to broad activity groups}
\usage{
caf2021_g2_groups()
}
\value{
A named list of integer vectors.
}
\description{
G2 has 73 categories. This function returns the default partition
used by [prepare_caf_data()]. Override by passing a list with the
same structure.
}
