% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jp_methods.R
\name{coef.jpfit}
\alias{coef.jpfit}
\title{Extract risk-function coefficients}
\usage{
\method{coef}{jpfit}(object, which = c("risk", "mean", "selection"), correction = TRUE, ...)
}
\arguments{
\item{object}{A `jpfit` object.}

\item{which}{One of "risk" (default), "mean", "selection". For "risk"
and "mean", choose `correction = TRUE/FALSE` to pick the with- or
without-IMR version.}

\item{correction}{Logical. Pick the corrected version when `TRUE`.}

\item{...}{Unused.}
}
\description{
Extract risk-function coefficients
}
